complete refactor to vite, typescript, new react updates
