+++ .prettierrc
@@ -1,1 +1,3 @@
-{}
+{
+    "tabWidth": 4
+}
