+++ src/components/PDFForm.tsx
@@ -8,13 +8,13 @@ const PDFForm = () => {
       <Row>
         <Col>
           <h2>PDF Forms</h2>
-          <CopyBlock
+          {/* <CopyBlock
             text={`<Form src={'https://examples.form.io/w4'} />`}
             theme={atomOneLight}
             language="jsx"
             codeBlock={true}
             showLineNumbers={false}
-          />
+          /> */}
           <div className="py-3">
             <Form src="https://examples.form.io/w4" />
           </div>
